backend/requirements.txt
flask==2.3.3
flask-cors==6.0.2
flask-sqlalchemy==3.0.5
flask-limiter==3.5.0
flask-jwt-extended==4.6.0
face-recognition==1.3.0
pillow==10.4.0
python-dotenv>=1.0.0
gunicorn==23.0.0
waitress==3.0.1

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
hypothesis==6.92.1
faker==20.1.0
